fix: add user info in signup mail
