+++ src/routes/inscription/_insertProfessionalAccount.gql
@@ -1,5 +1,10 @@
 mutation insertProfessionalAccount($account: account_insert_input!) {
 	account: insert_account_one(object: $account) {
 		id
+		professional {
+			structure {
+				name
+			}
+		}
 	}
 }
